Remove USE_SSE flag for rocksdb build
The `USE_SSE` flag explicitly enables the SSE4.2, PCLMUL, AVX2, BMI and
LZCNT instruction sets regardless of the host support, which may not be
desirable when targeting the legacy systems that do not support these
instructions.

This commit removes the `USE_SSE` flag so that the host compiler does
not emit any unsupported instructions.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>